% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate_relatedness.R
\name{estimate_rel}
\alias{estimate_rel}
\title{Estimate relatedness values}
\usage{
estimate_rel(data, input_allele_f = NULL)
}
\arguments{
\item{data}{A GENIND class object}

\item{input_allele_f}{A vector of allele frequencies (optional)}
}
\value{
A data.frame with \eqn{(n * (n - 1)) \ 2} (where \code{n} is the
         number of genotypes) rows of pairwise relatedness values for a
         number of relatedness indices.

 \itemize{
   \item{\code{ind1, ind2}: Pair of individuals for which relatedness values
       refer to.}
   \item{\code{n_missing_loci}: number of missing loci in the two individuals.}
   \item{\code{QG89_xy}: Queller and Goodnight (1989) index with \code{ind1} as the numerator and \code{ind2} as the denominator.}
   \item{\code{QG89_yx}: Queller and Goodnight (1989) index with \code{ind2} as the numerator and \code{ind1} as the denominator.}
   \item{\code{QG89_avg}: the mean of \code{QG89_xy} and \code{QG89_yx}.}
   \item{\code{QG89_rsxy}: sum of the numerators of \code{QG89_xy} and \code{QG89_yx} divided by the sum of the denominators of \code{QG89_xy} and \code{QG89_yx}.}
   \item{\code{LR99_avg}: the mean of the Lynch and Ritland (1999) index calculated with \code{ind1} as the reference individual, and  \code{ind2} as the reference.}
   \item{\code{W02_unc}: the uncorrected Wang (2002) index.}
   \item{\code{W02_cor}: the corrected Wang (2002) index.}
   \item{\code{prop_allele_shared}: proportion of alleles that are shared between \code{ind1} and \code{ind2}.}
   \item{\code{prop_loci_shared}: proportion of loci that are that have at least one allele in common between \code{ind1} and \code{ind2}.}
   \item{\code{missing_loci}: id of missing loci.}
 }
}
\description{
Estimate relatedness values
}
\details{
Queller and Goodnight (1989)

\deqn{r_{QG89_{xy}} = \sum_{l}\frac{(\sum_{alleles}(p_{xm} - \bar{p_{m}}))}{(\sum_{alleles}(p_{ym} - \bar{p_{m}}))}}{r_{QG89_{xy}} = sum_{l}((sum(p_xm - p_m)) / sum_{l}(sum(p_ym - p_m)))}

\deqn{r_{QG89_{yx}} = \sum_{l}\frac{(\sum_{alleles}(p_{ym} - \bar{p_{m}}))}{(\sum_{alleles}(p_{xm} - \bar{p_{m}}))}}{r_{QG89_{yx}} = sum_{l}((sum(p_ym - p_m)) / sum_{l}(sum(p_xm - p_m)))}
}
\examples{
library(adegenet)
data(nancycats)
estimate_rel(data = nancycats)
}
\references{
Queller and Goodnight. 1989.

Lynch and Ritland. 1999.
}

