% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irelr.R
\docType{package}
\name{irelr}
\alias{irelr}
\alias{irelr-package}
\title{irel: A package for reconstructing a pedigree using molecular data.}
\description{
The irelr package provides two important functions:
one that estimates relatedness, and one that simulated relatedness.
}
\section{Estimate relatedness}{

irelr provides a single function to estimate relatedness values for all 
pairs of individuals in a sample: irelr::estimate_rel()

Type ?irelr::estimate_rel() to find out more.
}

\section{Simulate relatedness}{

irelr provides a single function to simulated relatedness values for any
relatedness category: irelr::sim_rel()

Type ?irelr::sim_rel() to find out more
}

\section{Vignette}{

An extensive vignette is provided with examples of how to use irelr.
}

