% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/irelr.R
\name{sim_rel_dataframe}
\alias{sim_rel_dataframe}
\title{}
\usage{
sim_rel_dataframe(data, relCategories = c("po", "fs", "hs", "un"),
  numbIter = 10000, lociToKeep = "all", all_freqs = NULL)
}
\description{

}

